examples/multihop_rag/pyproject.toml
[project]
name = "multihop-rag"
version = "0.1.0"
description = "Multihop RAG Agent using GoalSeeking pattern with MultiHop-RAG benchmark"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "OpenSymbolicAI" }]
keywords = ["rag", "multihop", "goal-seeking", "agent", "llm", "opensymbolicai"]

dependencies = [
    "opensymbolicai-core",
    # Vector store
    "chromadb>=0.4.0",
    # Embeddings (Fireworks AI, OpenAI-compatible)
    "openai>=1.0.0",
    # HuggingFace datasets (MultiHop-RAG)
    "datasets>=2.0.0",
    # Environment variables
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
multihop-rag = "multihop_rag.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["multihop_rag"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
